requirements.txt
Flask==2.3.3
numpy==1.24.3
Pillow==10.0.1
scikit-learn==1.3.0
pandas==2.0.3